backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
openai==1.107.1
email-validator==2.2.0
python-multipart==0.0.20
pypdf==5.9.0
python-docx==1.2.0
